app/backend/requirements.txt
flask==3.0.3
python-dotenv==1.0.1
pdfplumber==0.11.4
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==11.1.0
rank-bm25==0.2.2
numpy==2.4.1

requests==2.32.3

schemdraw==0.19
